fix: container presentational pattern to dumper
