+++ src/app/[locale]/(dumper)/@contents/_add-contents-form/server.tsx
@@ -1,0 +1,8 @@
+import { hasDumperPostPermission } from "@/features/auth/utils/session";
+import { AddContentsFormClient } from "./client";
+
+export async function AddContentsForm() {
+	const hasPostPermission = await hasDumperPostPermission();
+	if (hasPostPermission) return <AddContentsFormClient />;
+	return <></>;
+}
