+++ src/app/[locale]/(dumper)/@contents/_contents-stack/server.tsx
@@ -1,25 +1,20 @@
-"use server";
-import "server-only";
 import { StatusCodeView } from "@/components/card/status-code-view";
 import { CardStack } from "@/components/stack/card-stack";
 import { getSelfId } from "@/features/auth/utils/session";
-import { deleteNews } from "@/features/news/actions/delete-news";
 import { loggerError } from "@/pino";
 import prisma from "@/prisma";
 
-export async function NewsStack() {
+export async function ContentsStack() {
 	try {
 		const userId = await getSelfId();
-
-		const unexportedNews = (
-			await prisma.news.findMany({
+		const unexportedContents = (
+			await prisma.contents.findMany({
 				where: { status: "UNEXPORTED", userId },
 				select: {
 					id: true,
 					title: true,
 					quote: true,
 					url: true,
-					Category: { select: { name: true } },
 				},
 				orderBy: { id: "desc" },
 			})
@@ -29,21 +24,15 @@ export async function NewsStack() {
 				title: d.title,
 				quote: d.quote,
 				url: d.url,
-				category: d.Category.name,
 			};
 		});
-		return (
-			<CardStack
-				data={unexportedNews}
-				deleteAction={deleteNews}
-				showDeleteButton
-			/>
-		);
+
+		return <CardStack data={unexportedContents} showDeleteButton={false} />;
 	} catch (error) {
 		loggerError(
 			"unexpected",
 			{
-				caller: "NewsStack",
+				caller: "ContentsStack",
 				status: 500,
 			},
 			error,
