+++ src/app/[locale]/(dumper)/@image/_add-image-form/server.tsx
@@ -1,0 +1,9 @@
+import { AddImageFormClient } from "@/app/[locale]/(dumper)/@image/_add-image-form/client";
+import { hasDumperPostPermission } from "@/features/auth/utils/session";
+
+export async function AddImageForm() {
+	const hasPostPermission = await hasDumperPostPermission();
+
+	if (!hasPostPermission) return <></>;
+	return <AddImageFormClient />;
+}
