--- src/app/[locale]/(dumper)/@image/_page.tsx
@@ -1,20 +1,0 @@
-import { Suspense } from "react";
-import { ImageStackSkeleton } from "@/components/stack/image-stack-skeleton";
-import { Separator } from "@/components/ui/separator";
-import { hasDumperPostPermission } from "@/features/auth/utils/session";
-import { AddImageForm } from "@/features/image/components/add-image-form";
-import { ImageStack } from "@/features/image/components/image-stack";
-
-export async function SuspensePage() {
-	const hasPostPermission = await hasDumperPostPermission();
-
-	return (
-		<>
-			{hasPostPermission && <AddImageForm />}
-			<Separator className="h-px bg-linear-to-r from-primary to-primary-grad" />
-			<Suspense fallback={<ImageStackSkeleton />}>
-				<ImageStack />
-			</Suspense>
-		</>
-	);
-}
