+++ src/app/[locale]/(dumper)/@image/components/all-image-stack.tsx
