+++ src/app/[locale]/(dumper)/@image/components/image-stack.tsx
