+++ src/app/[locale]/(dumper)/@image/page.tsx
@@ -1,6 +1,9 @@
 import { Suspense } from "react";
-import Loading from "@/components/loading";
-import { SuspensePage } from "./_page";
+import { ImageStack } from "@/app/[locale]/(dumper)/@image/components/image-stack";
+import { ImageStackSkeleton } from "@/components/stack/image-stack-skeleton";
+import { Separator } from "@/components/ui/separator";
+import { AddFormSkeleton } from "@/features/dump/components/add-form-skeleton";
+import { AddImageForm } from "./_add-image-form/server";
 
 type Props = {
 	searchParams: Promise<{ [key: string]: string | string[] | undefined }>;
@@ -15,8 +18,14 @@ export default async function Page({ searchParams }: Props) {
 	}
 
 	return (
-		<Suspense fallback={<Loading />}>
-			<SuspensePage />
-		</Suspense>
+		<>
+			<Suspense fallback={<AddFormSkeleton />}>
+				<AddImageForm />
+			</Suspense>
+			<Separator className="h-px bg-linear-to-r from-primary to-primary-grad" />
+			<Suspense fallback={<ImageStackSkeleton />}>
+				<ImageStack />
+			</Suspense>
+		</>
 	);
 }
