+++ src/app/[locale]/(dumper)/@news/_news-stack/server.tsx
@@ -1,22 +1,23 @@
-"use server";
-import "server-only";
 import { StatusCodeView } from "@/components/card/status-code-view";
 import { CardStack } from "@/components/stack/card-stack";
 import { getSelfId } from "@/features/auth/utils/session";
+import { deleteNews } from "@/features/news/actions/delete-news";
 import { loggerError } from "@/pino";
 import prisma from "@/prisma";
 
-export async function ContentsStack() {
+export async function NewsStack() {
 	try {
 		const userId = await getSelfId();
-		const unexportedContents = (
-			await prisma.contents.findMany({
+
+		const unexportedNews = (
+			await prisma.news.findMany({
 				where: { status: "UNEXPORTED", userId },
 				select: {
 					id: true,
 					title: true,
 					quote: true,
 					url: true,
+					Category: { select: { name: true } },
 				},
 				orderBy: { id: "desc" },
 			})
@@ -26,15 +27,21 @@ export async function ContentsStack() {
 				title: d.title,
 				quote: d.quote,
 				url: d.url,
+				category: d.Category.name,
 			};
 		});
-
-		return <CardStack data={unexportedContents} showDeleteButton={false} />;
+		return (
+			<CardStack
+				data={unexportedNews}
+				deleteAction={deleteNews}
+				showDeleteButton
+			/>
+		);
 	} catch (error) {
 		loggerError(
 			"unexpected",
 			{
-				caller: "ContentsStack",
+				caller: "NewsStack",
 				status: 500,
 			},
 			error,
