+++ src/app/[locale]/(dumper)/@news/page.tsx
@@ -1,6 +1,9 @@
 import { Suspense } from "react";
-import Loading from "@/components/loading";
-import { SuspensePage } from "./_page";
+import { NewsStack } from "@/app/[locale]/(dumper)/@news/_news-stack/server";
+import { CardStackSkeleton } from "@/components/stack/card-stack-skeleton";
+import { Separator } from "@/components/ui/separator";
+import { AddFormSkeleton } from "@/features/dump/components/add-form-skeleton";
+import { AddNewsForm } from "./_add-news-form/server";
 
 type Props = {
 	searchParams: Promise<{ [key: string]: string | string[] | undefined }>;
@@ -15,8 +18,16 @@ export default async function Page({ searchParams }: Props) {
 	}
 
 	return (
-		<Suspense fallback={<Loading />}>
-			<SuspensePage />
-		</Suspense>
+		<>
+			<Suspense fallback={<AddFormSkeleton showCategory />}>
+				<AddNewsForm />
+			</Suspense>
+
+			<Separator className="h-px bg-linear-to-r from-primary to-primary-grad" />
+
+			<Suspense fallback={<CardStackSkeleton />}>
+				<NewsStack />
+			</Suspense>
+		</>
 	);
 }
